==> 12S/Step7d_Permanova_eDNA_microDecon.Rmd <==

title: "Permanova_ZERO-impact"
author: "Isolde Cornelis & Laure Van den Bulcke"
date: "3/25/2020"
output: html_document

------
INPUT 
------

Data inlezen. 
Ons experimental design bestaat uit 3 niveaus:
- Impact
- biologische replicaten
- DNA replicaten
- van 1 DNA replicaat per biologisch replicaat werden 3 PCR replicaten genomen. Maar aangezien dit geen fully nested design is (= van elk DNA replicaat 3 PCR replicaten), moeten we 2x een permanova uitvoeren: een op de data van de DNA replicaten en een op de data van de PCR replicaten.
LET OP! 
We kunnen hierdoor kijken naar het relatief belang van Impacts, biologische replicaten en DNA replicaten enerzijds en naar het relatief belang van Impacts, biologische replicaten en PCR replicaten anderzijds, maar we kunnen DNA en PCR replicaten onderling niet vergelijken !

Daarnaast is het belangrijk om in rekening te houden dat de resultaten maar gebaseerd zijn op 3 Impacts (dus geen weergave van BDNS), besluiten mag je dus enkel nemen op deze Impacts en niet veralgemenen. Voorzichtig zijn in verwoording van de resultaten!


Beta diversity: only look at rarefied data 
```{r Read data}
if(.Platform$OS.type == "unix"){home<-"/home/"} else{
  home<-"//192.168.236.131/"}

# make paths
proj.path <- here("/home/genomics/icornelis/02_ZEROimpact/01_12S/NJ2021/MiFish-UE_run2")

#1 read data
#setwd(paste0("home/genomics/icornelis/Demultiplex_Liu"))
#upload data
table_unrarefied <- readxl::read_excel(paste0(proj.path,"/MiFish_UE-S_concatenated/results_microDecon/table_unrarefied_concatenated_FullTaxonomicAssignment_clean.xlsx"))
env <- readRDS(paste0(proj.path,"/MiFish_UE-S_concatenated/results_microDecon/R_Environment/env_ordered_noNeg.rds"))
table_unrarefied <- data.frame((table_unrarefied), check.names = FALSE)

#remove samples from environmental data that are absent from the rarefied data
keep_samples <- c(colnames(table_unrarefied))
env_unrarefied <- env[env$Niskin.sample %in% keep_samples,]
colorder <- c(env_unrarefied$Niskin.sample)

#select ASVs that are assigned to Fish
fish_classes <- readRDS(file = paste0(proj.path,"/MiFish_UE-S_concatenated/results_v2/REnvironment/Fish_classes.rds"))
freshwater_fish <- readRDS(file = paste0(proj.path,"/MiFish_UE-S_concatenated/results_v2/REnvironment/Fish_Freshwater.rds"))
table_unrarefied_FishASVs <- as.data.frame(table_unrarefied[table_unrarefied$Class %in% fish_classes,])
table_unrarefied_FishASVs <- as.data.frame(table_unrarefied_FishASVs[!table_unrarefied_FishASVs$Species %in% c(freshwater_fish, "NA"),])
table_unrarefied_FishASVs[is.na(table_unrarefied_FishASVs)] <- 0

#3: sort by species
taxo <- "Species"
merged_data_unrarefied_fish <- aggregate(table_unrarefied_FishASVs[,1:(ncol(table_unrarefied_FishASVs)-11)], by= list(as.factor(table_unrarefied_FishASVs[,taxo])),FUN=sum)
rownames(merged_data_unrarefied_fish) <- merged_data_unrarefied_fish$Group.1
merged_data_unrarefied_fish$Group.1 <- NULL
Fish_zero <- c(rownames(merged_data_unrarefied_fish[rowSums(merged_data_unrarefied_fish) == 0,]))
Fish_empty <- c(colnames(merged_data_unrarefied_fish[colSums(merged_data_unrarefied_fish) == 0]))
merged_data_unrarefied_fish <- merged_data_unrarefied_fish[!rownames(merged_data_unrarefied_fish) %in% Fish_zero,]
merged_data_unrarefied_fish <- merged_data_unrarefied_fish[,!colnames(merged_data_unrarefied_fish) %in% Fish_empty,]

env <- env[!env$Niskin.sample %in% Fish_zero,]

```


De nodige dataframes worden gemaakt uit de ingelezen data:
1) informatie over de soorten: rijen= stalen, kolommen= soortsnamen
2) extra informatie: rijen= stalen, kolommen= Impact, biologische & DNA replicaten

```{r Needed dataframes, include=FALSE}
library(vegan)
#data_soorten <- as.data.frame(t(table_unrarefied_FishASVs[,1:(ncol(table_unrarefied_FishASVs)-11)]))
data_soorten <- as.data.frame(t(merged_data_unrarefied_fish))

keep_samples <- c(rownames(data_soorten))
env_unrarefied <- env[env$Niskin.sample %in% keep_samples,]

data_soorten <- data_soorten %>% filter(!grepl("neg", rownames(data_soorten)))
env <- env_unrarefied %>% filter(!grepl("neg", env_unrarefied$Niskin.sample))

replicaten_name <- rownames(data_soorten)

rownames(data_soorten) <- env$Niskin.sample
data_replicaten<- as.data.frame(rownames(data_soorten))
colnames(data_replicaten)<-"V1"
data_replicaten$Impact <- env$Location
data_replicaten$Biol_replicate <- env$Biol_replicate
data_replicaten$Impact_biol <- paste(data_replicaten$Impact,data_replicaten$Biol_replicate,sep="_")
data_replicaten$Environment <- env$Environment
data_replicaten$Zone <- env$Zone

data_replicaten$Impact <- as.factor(data_replicaten$Impact)
data_replicaten$Impact_biol <- as.factor(data_replicaten$Impact_biol)
data_replicaten$Biol_replicate <- as.factor(data_replicaten$Biol_replicate)
data_replicaten$Environment <- as.factor(data_replicaten$Environment)
data_replicaten$Zone <- as.factor(data_replicaten$Zone)
data_replicaten$Env_Zon <- as.factor(paste(data_replicaten$Environment,data_replicaten$Zone,sep="-"))

data_replicaten_OWF <- data_replicaten %>% filter(grepl("_OWF", data_replicaten$Environment))
data_soorten_OWF <- data_soorten[data_replicaten_OWF$V1,]

library("xlsx")
write.xlsx(data_soorten, paste0(proj.path,"/Step5_Statistics/data_soorten.xlsx"), sheetName = "Sheet1", col.names = TRUE, row.names = TRUE, append = FALSE)
write.xlsx(data_replicaten, paste0(proj.path,"/Step5_Statistics/data_replicaten.xlsx"), sheetName = "Sheet1", col.names = TRUE, row.names = FALSE, append = FALSE)

```



Een Dissimilarity index wordt berekend. Hier Jaccard (aanwezigheid/afwezigheid)

```{R Dissimilarity index }
# JACCARD

#dissimilarity index = jaccard (prensence/absence)
dist_soorten_jaccard <- vegdist(data_soorten, method = "jaccard",binary=F) 

# BRAY
range(data_soorten)
dist_soorten_bray <- vegdist(data_soorten, method = "bray", binary = F) 
dist_soorten_OWF_bray <- vegdist(data_soorten_OWF, method ="bray", binary = F)

```
Ons experimental design bestaat uit 2 niveaus:
- Zone
- Impact
Het laagste niveau = residuals

H0: „the centroids of the groups, as defined in the space of the chosen resemblance measure, are equivalent for all groups.“. Thus, if H0 were true, any observed differences among the centroids in a given set of data will be similar in size to what would be obtained under random allocation of individual sample units to the groups (i.e., under permuration).

NESTING 
This answer is based on the adonis function in the vegan package for R. Nesting can enter a permanova in two ways: your model and your permutation structure (the strata parameter).

In your model formula, nesting is simply interaction as specified in R with the forward slash or colon (A/B/C or A:B:C); you are specifying that the effect of C should differ based on the value of B and A. This is like a fixed effect. Also note that forward slash differs from colon in that it includes terms for all layers in the nesting hierarchy: A+A:B+A:B:C.

With strata=A:B, you limit your permutations to be between only those samples with the same levels of A and B. This implicitly treats data from each A:B interaction level as coming from the separate distributions/having disparate error structure. It is therefore somewhat similar to a random effect.

```{R Permanova}

#Transform the data for normal distribution (choose exponent through box-cox transformation test)
#data_soorten_transformed <- decostand(data_soorten, method="log")
#data_soorten_transformed <- decostand(data_soorten, method="pa")
#data_soorten_transformed <- (data_soorten)^0.25
#data_soorten_transformed <- decostand(data_soorten, method="hellinger")
data_soorten_transformed <- decostand(decostand(data_soorten, method="tot"), method="max")
dist_soorten_transformed_bray <- vegdist(data_soorten_transformed, method = "bray", binary = F) 


#data_soorten_OWF_transformed <- decostand(data_soorten_OWF, method="log")
#data_soorten_OWF_transformed <- decostand(data_soorten_OWF, method="pa")
#data_soorten_OWF_transformed <- (data_soorten_OWF)^0.25
#data_soorten_OWF_transformed <- decostand(data_soorten_OWF, method="hellinger")
data_soorten_OWF_transformed <- decostand(decostand(data_soorten_OWF, method="tot"), method="max")
dist_soorten_OWF_transformed_bray <- vegdist(data_soorten_OWF_transformed, method = "bray", binary = F) 

Impact_p <- data_replicaten$Environment
Zone_p <- data_replicaten$Zone
Impact_OWF_p <- data_replicaten_OWF$Environment
Zone_OWF_p <- data_replicaten_OWF$Zone
Zone_Impact_OWF_p <- data_replicaten_OWF$Env_Zon

Impact_biol_repl_p <- data_replicaten$Impact_biol
Biol_repl_p <- data_replicaten$Biol_replicate
Biol_repl_OWF_p <- data_replicaten_OWF$Biol_replicate


#ASSUMPTIONS

library(pairwiseAdonis)
pairwise.adonis(dist_soorten_transformed_bray, Impact_p, p.adjust.m="bonferroni", reduce=NULL, perm=9999)
pairwise.adonis(dist_soorten_transformed_bray, Zone_p, p.adjust.m="bonferroni", reduce=NULL, perm=9999)
pairwise.adonis(dist_soorten_OWF_transformed_bray, Impact_OWF_p, p.adjust.m="bonferroni", reduce=NULL, perm=9999)
pairwise.adonis(dist_soorten_OWF_transformed_bray, Zone_OWF_p, p.adjust.m="bonferroni", reduce=NULL, perm=9999)
pairwise.adonis(dist_soorten_OWF_transformed_bray, Zone_Impact_OWF_p, p.adjust.m="bonferroni", reduce=NULL, perm=9999)
#geen meerwaard

library(pairwiseAdonis)
pairwise.adonis(dist_soorten_bray, Impact_p, p.adjust.m="BH", reduce=NULL, perm=9999)
pairwise.adonis(dist_soorten_bray, Zone_p, p.adjust.m="BH", reduce=NULL, perm=9999)
pairwise.adonis(dist_soorten_OWF_bray, Impact_OWF_p, p.adjust.m="BH", reduce=NULL, perm=9999)
pairwise.adonis(dist_soorten_OWF_bray, Zone_OWF_p, p.adjust.m="BH", reduce=NULL, perm=9999)
pairwise.adonis(dist_soorten_OWF_bray, data_replicaten_OWF$Env_Zon, p.adjust.m="BH", reduce=NULL, perm=9999)

```
Permanova significant voor Zone en Impact (residuals).

A significant Permanova means one of three things. 1) There is a difference in the location of the samples (i.e. the average community composition), 2) There is a difference in the dispersion of the samples (i.e. the variability in the community composition), or 3) There is a difference in both the location and the dispersion.
So, if you get a significant Permanova you'll want to distinguish between the three options. That of course is why you need to run the permdisp. If you get a non-significant Permdisp you can conclude the first option above is the correct one. If you get a significant Permdisp then it is either the second or third option (that is there is definitely a difference in dispersion and maybe a difference in location). There is no fool-proof way to distinguish between these two, but looking at an MDS plot of the data will hopefully help you do so (also see note 2 below).

ASSUMPTIONS

In the case of a one-way analysis, the PERMANOVA test using
permutations assumes only that the samples are exchangeable under a true null hypothesis

PERMANOVA works with the assumption that dispersion of the data in your samples is the ~same among each other, so before running PERMANOVA, you must run the betadisper->permutest to know if the dispersions are the same. For that to be true, the permutest has to have a non-significant p-value. Knowing the previous, then you can run the PERMANOVA test, otherwise your interpretations will be wrong
=> dispersion testen met permdisp


```{R, PERMANOVA Bray}

# Advanced applied statistieks

pmv_test_bray <- adonis2(dist_soorten_bray ~   Zone_p, data=data_replicaten, permutations=9999)
pmv_test_bray
pmv_OWF_bray <- adonis2(dist_soorten_OWF_bray ~ Zone_OWF_p * Impact_OWF_p * Zone_OWF_p/Impact_OWF_p, data=data_replicaten_OWF, permutations = 9999)
pmv_OWF_bray

#factor Zone
Zone_betadisper_j <- betadisper(dist_soorten_bray, Zone_p, type="centroid")
Zone_permdisp_j <- permutest(Zone_betadisper_j, permutations=9999)
Zone_permdisp_j #SIGN
plot(Zone_betadisper_j, hull=FALSE, ellipse=TRUE)

#factor Zone OWF
OWF_betadisper_j <- betadisper(dist_soorten_OWF_bray, Zone_Impact_OWF_p, type="centroid")
OWF_permdisp_j <- permutest(OWF_betadisper_j, permutations=9999)
OWF_permdisp_j # SIGN
plot(OWF_betadisper_j, hull=FALSE, ellipse=TRUE)

#log root transformation
pmv_test_bray_Zone_transformed <- adonis2(dist_soorten_transformed_bray ~   Zone_p, data=data_replicaten, permutations=9999)
pmv_test_bray_Zone_transformed
pmv_OWF_bray_transformed <- adonis2(dist_soorten_OWF_transformed_bray~ Zone_OWF_p * Impact_OWF_p * Zone_OWF_p/Impact_OWF_p, data=data_replicaten_OWF, permutations = 9999)
pmv_OWF_bray_transformed

#factor Zone
Transformed_Zone_betadisper_j <- betadisper(dist_soorten_transformed_bray, Zone_p, type="centroid")
Transformed_Zone_permdisp_j <- permutest(Transformed_Zone_betadisper_j, permutations=9999)
Transformed_Zone_permdisp_j #SIGN
plot(Transformed_Zone_betadisper_j, hull=FALSE, ellipse=TRUE)

#factor Zone OWF
Transformed_OWF_betadisper_j <- betadisper(dist_soorten_OWF_transformed_bray, Zone_Impact_OWF_p, type="centroid")
Transformed_OWF_betadisper_zone <- betadisper(dist_soorten_OWF_transformed_bray, Zone_OWF_p, type="centroid")
Transformed_OWF_betadisper_impact <- betadisper(dist_soorten_OWF_transformed_bray, Impact_OWF_p, type="centroid")
Transformed_OWF_permdisp_j <- permutest(Transformed_OWF_betadisper_j, permutations=9999)
Transformed_OWF_permdisp_z <- permutest(Transformed_OWF_betadisper_zone, permutations=9999)
Transformed_OWF_permdisp_i <- permutest(Transformed_OWF_betadisper_impact, permutations=9999)
Transformed_OWF_permdisp_j 
Transformed_OWF_permdisp_z
Transformed_OWF_permdisp_i

plot(Transformed_OWF_betadisper_j, hull=FALSE, ellipse=TRUE, label = T)

library("car")
qqPlot(Transformed_OWF_betadisper_j$distances)
qqPlot(Transformed_Zone_permdisp_j$distances)
```
